+++ addon/pico/sample/kernel.cpp
@@ -1,0 +1,104 @@
+//
+// kernel.cpp
+//
+// Circle - A C++ bare metal environment for Raspberry Pi
+// Copyright (C) 2014-2021  R. Stange <rsta2@o2online.de>
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//
+#include "kernel.h"
+
+// GPIO pin configuration
+#define SWCLK_PIN		17
+#define SWDIO_PIN		27
+#define RESET_PIN		24		// 0 for none
+
+#define SWD_CLOCK_RATE_KHZ	400
+
+#define RP2040_RAM_BASE		0x20000000U
+
+static const u8 Firmware[] =
+{
+#include "firmware.h"
+};
+
+LOGMODULE ("kernel");
+
+CKernel::CKernel (void)
+:	m_Screen (m_Options.GetWidth (), m_Options.GetHeight ()),
+	m_Timer (&m_Interrupt),
+	m_Logger (m_Options.GetLogLevel (), &m_Timer),
+	m_FirmwareLoader (SWCLK_PIN, SWDIO_PIN, RESET_PIN, SWD_CLOCK_RATE_KHZ)
+{
+	m_ActLED.Blink (5);	// show we are alive
+}
+
+CKernel::~CKernel (void)
+{
+}
+
+boolean CKernel::Initialize (void)
+{
+	boolean bOK = TRUE;
+
+	if (bOK)
+	{
+		bOK = m_Screen.Initialize ();
+	}
+
+	if (bOK)
+	{
+		bOK = m_Serial.Initialize (115200);
+	}
+
+	if (bOK)
+	{
+		CDevice *pTarget = m_DeviceNameService.GetDevice (m_Options.GetLogDevice (), FALSE);
+		if (pTarget == 0)
+		{
+			pTarget = &m_Screen;
+		}
+
+		bOK = m_Logger.Initialize (pTarget);
+	}
+
+	if (bOK)
+	{
+		bOK = m_Interrupt.Initialize ();
+	}
+
+	if (bOK)
+	{
+		bOK = m_Timer.Initialize ();
+	}
+
+	if (bOK)
+	{
+		bOK = m_FirmwareLoader.Initialize ();
+	}
+
+	return bOK;
+}
+
+TShutdownMode CKernel::Run (void)
+{
+	LOGNOTE ("Compile time: " __DATE__ " " __TIME__);
+
+	if (!m_FirmwareLoader.Load (Firmware, sizeof Firmware, RP2040_RAM_BASE))
+	{
+		LOGPANIC ("Firmware load failed");
+	}
+
+	return ShutdownHalt;
+}