+++ include/circle/machineinfo.h
@@ -36,6 +36,7 @@ enum TMachineModel
 	MachineModelBPlus,
 	MachineModelZero,
 	MachineModelZeroW,
+	MachineModelZero2W,
 	MachineModel2B,
 	MachineModel3B,
 	MachineModel3APlus,