+++ lib/machineinfo.cpp
@@ -73,6 +73,7 @@ s_NewInfo[]
 	{14, MachineModel3APlus,	3},
 	{16, MachineModelCM3Plus,	3},
 	{17, MachineModel4B,		4},
+	{18, MachineModelZero2W,	3},
 	{19, MachineModel400,		4},
 	{20, MachineModelCM4,		4}
 };
@@ -87,6 +88,7 @@ static const char *s_MachineName[] =		// must match TMachineModel
 	"Raspberry Pi Model B+",
 	"Raspberry Pi Zero",
 	"Raspberry Pi Zero W",
+	"Raspberry Pi Zero 2 W",
 	"Raspberry Pi 2 Model B",
 	"Raspberry Pi 3 Model B",
 	"Raspberry Pi 3 Model A+",
@@ -119,6 +121,7 @@ static unsigned s_ActLEDInfo[] =		// must match TMachineModel
 	47,				// B+
 	47 | ACTLED_ACTIVE_LOW,		// Zero
 	47 | ACTLED_ACTIVE_LOW,		// Zero W
+	29 | ACTLED_ACTIVE_LOW,		// Zero 2 W
 	47,				// 2B
 	0 | ACTLED_VIRTUAL_PIN,		// 3B
 	29,				// 3A+
@@ -359,7 +362,8 @@ unsigned CMachineInfo::GetGPIOPin (TGPIOVirtualPin Pin) const
 	case GPIOPinAudioLeft:
 #ifdef USE_PWM_AUDIO_ON_ZERO
 		if (   m_MachineModel == MachineModelZero
-		    || m_MachineModel == MachineModelZeroW)
+		    || m_MachineModel == MachineModelZeroW
+		    || m_MachineModel == MachineModelZero2W)
 		{
 #ifdef USE_GPIO18_FOR_LEFT_PWM_ON_ZERO
 			return 18;
@@ -388,7 +392,8 @@ unsigned CMachineInfo::GetGPIOPin (TGPIOVirtualPin Pin) const
 	case GPIOPinAudioRight:
 #ifdef USE_PWM_AUDIO_ON_ZERO
 		if (   m_MachineModel == MachineModelZero
-		    || m_MachineModel == MachineModelZeroW)
+		    || m_MachineModel == MachineModelZeroW
+		    || m_MachineModel == MachineModelZero2W)
 		{
 #ifdef USE_GPIO19_FOR_RIGHT_PWM_ON_ZERO
 			return 19;
@@ -478,7 +483,8 @@ boolean CMachineInfo::ArePWMChannelsSwapped (void) const
 {
 	return    m_MachineModel >= MachineModelAPlus
 	       && m_MachineModel != MachineModelZero
-	       && m_MachineModel != MachineModelZeroW;
+	       && m_MachineModel != MachineModelZeroW
+	       && m_MachineModel != MachineModelZero2W;
 }
 
 unsigned CMachineInfo::AllocateDMAChannel (unsigned nChannel)