+++ lib/net/ntpdaemon.cpp
@@ -2,7 +2,7 @@
 // ntpdaemon.cpp
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2015-2020  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2015-2021  R. Stange <rsta2@o2online.de>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -33,6 +33,8 @@ CNTPDaemon::CNTPDaemon (const char *pNTPServer, CNetSubSystem *pNetSubSystem)
 	m_pNetSubSystem (pNetSubSystem)
 {
 	assert (m_pNetSubSystem != 0);
+
+	SetName (FromNTPDaemon);
 }
 
 CNTPDaemon::~CNTPDaemon (void)