+++ lib/net/phytask.cpp
@@ -2,7 +2,7 @@
 // phytask.h
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2019  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2019-2021  R. Stange <rsta2@o2online.de>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -24,6 +24,7 @@
 CPHYTask::CPHYTask (CNetDevice *pDevice)
 :	m_pDevice (pDevice)
 {
+	SetName ("netphy");
 }
 
 CPHYTask::~CPHYTask (void)